requirements.txt
streamlit
pydantic
pillow
